requirements.txt
# Core Dependencies
web3==6.15.1
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.31.0

# Database
supabase==2.3.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Async Support
aiohttp==3.9.3
aiofiles==23.2.1

# Alert Systems
python-telegram-bot==20.8
smtplib-ssl==1.0.0
discord-webhook==1.3.1

# Pattern Detection & ML
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
scipy==1.12.0

# GraphQL for Bitquery
gql==3.5.0
graphql-core==3.2.3
aiohttp[speedups]==3.9.3

# Monitoring & Logging
python-json-logger==2.0.7
coloredlogs==15.0.1
rich==13.7.0

# Utils
python-dateutil==2.8.2
pytz==2024.1
click==8.1.7
tabulate==0.9.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Rate Limiting
ratelimit==2.2.1

# Caching
cachetools==5.3.2
redis==5.0.1

# HTTP & API
httpx==0.26.0
tenacity==8.2.3

# Cryptography
eth-account==0.11.0
eth-utils==2.3.1
eth-abi==4.2.1

# Configuration
hydra-core==1.3.2
omegaconf==2.3.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
ipython==8.20.0